Move webhook endpoint tests to specific folder
