+++ tests/controllers/auth/webhook/register.test.ts
@@ -1,20 +1,20 @@
 import getMockedClient, {
   AuthClientMock,
-  onResetConfirmOutput,
+  onRegisterUploadOutput,
 } from '@mocks/lib/auth';
 jest.mock('@lib/auth', () => getMockedClient);
 
 import express from 'express';
 import request from 'supertest';
-import { onResetConfirm } from '../../../src/controllers/auth/webhook/resetConfirm';
-import { expectResolved } from '../../utils';
+import { onRegister } from '@controllers/auth/webhook/register';
+import { expectResolved } from '../../../utils';
 
-describe('Reset Confirm Webhook Endpoint', () => {
+describe('Register Webhook Endpoint', () => {
   const app = express();
 
   beforeAll(() => {
     app.use(express.json({ type: 'application/json' }));
-    app.post('/', onResetConfirm);
+    app.post('/', onRegister);
   });
   afterAll(() => {
     jest.restoreAllMocks();
@@ -29,7 +29,7 @@ describe('Reset Confirm Webhook Endpoint', () => {
     getMockedClient.mockImplementationOnce(() => {
       throw new Error('Intentional get auth client error');
     });
-    AuthClientMock.onResetConfirm.mockImplementationOnce(() => {
+    AuthClientMock.onRegisterUpload.mockImplementationOnce(() => {
       throw new Error('Intentional onRegisterUpload error');
     });
     const responses = await Promise.all([
@@ -45,11 +45,13 @@ describe('Reset Confirm Webhook Endpoint', () => {
     const response = await request(app).post('/').send({ fake: 'data' });
     expect(getMockedClient).toHaveBeenCalled();
     expect(getMockedClient).toHaveReturnedWith(AuthClientMock);
-    expect(AuthClientMock.onResetConfirm).toHaveBeenCalled();
-    expectResolved(AuthClientMock.onResetConfirm).toEqual(onResetConfirmOutput);
-    expect(response.ok);
-    expect(response.body).toMatchObject({
-      EACResetToken: onResetConfirmOutput,
+    expect(AuthClientMock.onRegisterUpload).toHaveBeenCalledWith({
+      fake: 'data',
     });
+    expectResolved(AuthClientMock.onRegisterUpload).toMatchObject(
+      onRegisterUploadOutput
+    );
+    expect(response.ok);
+    expect(response.body).toMatchObject(onRegisterUploadOutput);
   });
 });
